fail CI if coverage fails to upload (#5101)
**Context:**
When the Codecov upload action fails because of network issues or the
like, the action reports as passed, then the `codecov` action stays as
"Expected" forevermore. This is worse than a failure because it
continues to block us but we don't immediately know why. If we knew, we
could just re-run it when we find out.

**Description of the Change:**
Set the codecov upload action to fail on failure 😄 

**Benefits:**
We will know when it fails to upload, and we can re-run the upload task
manually

**Possible Drawbacks:**
External contributors might hit this, and although the old ways were
also bad, this might be a bit more visibly frightening if you don't know
why you got a big red ❌