Qutrit mixed apply operation (#5032)
**Context:**
Currently the qutrit_mixed device is being developed this is a necessary
addition to the overall project as this applies operations to a qutrit
mixed state. This is a prerequisite for other functionality relating to
the qutrit mixed device for noisy qutrit simulation.

**Description of the Change:**
Added functionality for applying operations to a qutrit mixed-state. The
new ``apply_operation`` function can be used to apply gates and Channels
to a qutrit mixed-state.

**Benefits:**
Allows for Channels and operations to be applied to a mixed state, will
be used to add execute functionality to qutrit mixed-state device
allowing for noisy qutrit simulation

**Possible Drawbacks:**
Abstracting for qubits and more generally qutrits may have added
challenges and will require a reforctor or copied code code-smell.

**Related GitHub Issues:**
N/A

---------

Co-authored-by: Gabe PC <bottrill@student.ubc.ca>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Christina Lee <chrissie.c.l@gmail.com>
Co-authored-by: Olivia Di Matteo <2068515+glassnotes@users.noreply.github.com>