Import state sign convention change and test improvements (#5114)
**Context:**

Initial state preparation in PennyLane now imports states based on the
physicist sign convention (interleaved operators). This allows exact
agreement when energies are evaluated for the same state using
PennyLane's Hamiltonian matrix, leading to much better acceleration of
quantum algorithms.

**Description of the Change:**

A sign convention is applied to initial states that originated in PySCF
(CISD/CCSD) or Dice (SHCI) to convert them to physicist convention in
PennyLane (interleaved spin-up / spin-down operators). Additionally, new
tests are introduced comparing the energy of the initial state from the
method which contributed the state, to that same state's energy as
evaluated in PennyLane. This is an extra check ensuring that the state
import process is correct.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**