Add grouping support for arithmetic ops (#5070)
**Context:**
Adding support for `Prod` and `SProd` to be used with grouping.

**Description of the Change:**
* Update `Tensor.__matmul__` to stop mutating in-place and return a new
op instead.
* Added Pauli rep to `Tensor`. Consequently, `Tensor.copy()` and
`Tensor.map_wires()` also had to be updated.
* Moved around code inside `Tensor__init__` to fix incorrect validation
after the previous changes. The warning about overlapping wires is now
raised in `Tensor.queue()` instead of at the beginning of construction,
so that raising the warning pre-maturely can be prevented.
* `qml.pauli.group_observables` now returns `Prod`s if the observable
list contains new arithmetic ops.
* Update `pauli_to_binary` to rely on the Pauli rep. All operations that
can be grouped should now always have a valid Pauli rep, so we don't
need to worry about other cases, as `pauli_to_binary` is only a utility
function for grouping.
* Update `are_pauli_words_qwc` to correctly handle ops with Pauli reps.
This change however means that any new arithmetic ops that have
`pauli_rep=None` will still not work with this function. If we
eventually start doing more simplification eagerly, this can be
addressed.
* `split_non_commuting` now creates `Prod`s instead of `Tensor`s for
creating observables for measurements that use wires. Also made a change
so that if a measurement without any wires is made, the observable that
is created uses `tape.wires`.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: qottmann <korbinian.kottmann@gmail.com>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>