[Terms1] Add `Prod.terms()` method (#5132)
The overall goal is to have `Sum.terms()` give meaningful results. This
is necessary to achieve feature parity with `qml.Hamiltonian` to be able
to soon switch to new opmath be the default. We realized that a
necessity for `Sum.terms()` to work properly we also need a
`Prod.terms()` method that is compatible with that.

In particular with this PR you can do:

```
>>> qml.operation.enable_new_opmath()
>>> op = X(0) @ (0.5 * X(1) + X(2))
>>> op.terms()
([0.5, 1.0],
 [PauliX(wires=[1]) @ PauliX(wires=[0]),
  PauliX(wires=[2]) @ PauliX(wires=[0])])
```

You can think of it like this: When the operator has a `pauli_rep` its
just taking all values as coeffs and all keys as operators. This PR
additionally handles all cases where no pauli rep is available. In
essence, it is doing the same steps as `simplify` modulo some small
differences (like not discerning cases of just one single term as that
still yields `[factor], [op]`).

- [x] Basic functionality
- [x] Tests with Paulis
- [x] Tests without Paulis
- [x] Docs
- [x] changelog

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>