Change naming convention for compilers' entry points. (#5140)
**Context:** We would like to:

* Allow package publishers to provide multiple compiler implementations
within a single package.
* Allow users to use different compiler implementations within a single
package via (possibly but not necessarily) the `compiler` keyword.

**Description of the Change:** We obtain the name of the compiler from
the entry's name. This allows multiple compilers to be defined in a
single package. [See
here](https://github.com/PennyLaneAI/catalyst/pull/493) for the
implementation in Catalyst.

**Benefits:** The current use case is for the Catalyst package to
provide support for the `catalyst` compiler and a `cudaq` based compiler
and allow the user to choose this compiler with the syntax:
`qml.qjit(compiler="cudaq")`.

**Possible Drawbacks:** adds a little more work to package maintainers.
They now have to prepend the name of the compilers to their entry
points.

**Related GitHub Issues:**

[sc-55975]

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>