`qml.dot` always returns a `Sum` when there's more than one term. (#5143)
[sc-55330]

See also PR #5073 and Issue #5066 . Sometimes `qml.dot` was returning a
scalar product between a single coefficient and a sum, instead of
returning a sum. `TrotterProduct` was not able to handle this sort of
nested operator.

While we should also make our code more flexible in handling nested
operators, we can also standardize the return of `qml.dot`.