CI: fetch and check out master using checkout action (#5204)
it
[failed](https://github.com/PennyLaneAI/pennylane/actions/runs/7904108221/job/21574971597)
on master again while trying to merge. truth is, we just need to set
things up correctly and then not mess with master.

I deleted the `rm` call because `download-artifact` will just overwrite
it anyway, and on the off-chance the set of artifacts changes on master
for some reason, I want that diff to be caught.