TransformDispatcher can dispatch onto a batch of tapes (#5163)
**Context:** A primary goal for the transform project is that transforms
are composable. With the current features, transforms are composable on
the qnode, but they need to be more robust to compose when working with
(batches of) tapes.

**Description of the Change:** A new private method has been added to
the `TransformDispatcher` class to dispatch transforms onto batches of
tapes. Such a method is essentially the copy of `map_batch_transform`.

**Benefits:** The proposed solution would lower cognitive overhead for
developers working with transforms.

**Possible Drawbacks:** The new method is called only if the class
object is a sequence with all instances of `qml.tape.QuantumScript`.
Otherwise, a previously implemented `TransformError` is raised.

**Related GitHub Issues:** None.

---------

Co-authored-by: Christina Lee <chrissie.c.l@gmail.com>