make CompositeOp.eigendecomposition jit-compatible (#5207)
**Context:**
Similar to #5206 - `state_diagonalizing_gates` depends on the
eigendecomposition which is not jit-compatible.

**Description of the Change:**
use `qml.math` throughout eigval computation in order for it to be
jit-compatible

**Benefits:**
Anything depending on the eigendecomposition (eg. measuring an
observable with abstract data) will now work with jit

**Possible Drawbacks:**
The solution in #5206 might be more performant? I have closed that PR,
but pls lmk if you think I should re-open it.

[sc-56801]