fix qubitdensitymatrix on default.mixed (#5203)
**Context:**
Jax-jit (or any tracing tool) does not allow reading/depending on the
concrete value of a tracer. `default.mixed` currently validates the
trace of the parameter passed to `QubitDensityMatrix` before applying
it.

**Description of the Change:**
If the parameter is abstract, do not validate it

**Benefits:**
`default.mixed` can now use `QubitDensityMatrix` with jax-jit

**Possible Drawbacks:**
It will now try to apply invalid density matrices because it cannot
check it. This (skipping validation of concrete values while tracing) is
common practice in PennyLane, so it should be ok.

**Related GitHub Issues:**
Fixes #5196