Native mid-circuit measurements support composite measurements, etc. (phase 2) (#5120)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
This PR is a follow-up on #5088 .

[Epic 50811](https://app.shortcut.com/xanaduai/epic/50811)  states:

- [x] The deferred measurement principle is not used
- [x] Each shot is a new exploration of the tree of possible mid-circuit
measurement results
- [x] The mid-circuit measurement results are correctly exposed to
PennyLane's frontend
- [x] Compatible with qubit reuse
- [x] Compatible with qubit reset
- [x] Compatible with qubit postselection
- [x] Compatible with existing support for conditioning in PennyLane
- [x] Addition, subtraction and multiplication of one or more
mid-circuit measurement results

The current PR tackles the requirements above as indicated by the task
list.

**Description of the Change:**

- `apply_mid_measure` handles post-selection returning a zero state and
sample when the state isn't properly normalized (which happens when the
state cannot be normalized).
- The `concretize` method is added to MeasurementValue object to obtain
a concrete value from a dictionary of measurement values. This allows
handling composite measurements. Other modules like `simulate` are
simplified accordingly.
- Post-selected-zero-state samples are thrown away at the `simulate`
level, handling this exception. This means the final number of samples
is lower or equal to the total number of shots.
- Tests for all introduced features.

**Benefits:**
Mid-circuit measurements are natively supported.

**Possible Drawbacks:**
Slower in several cases (e.g. when few MCMs).

**Missing features:**
To do in follow-up PRs:

- [ ] parameter-shift support
- [ ] optimisation

**Related GitHub Issues:**
[sc-54987]
[sc-55666]
[sc-55667]
[sc-55668]
[sc-55922]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>