start to freeze dependencies for docs (#5228)
**Context:**
Now that we have a workflow that freezes deps when running tests, we
should have one that freezes deps when building docs too! This is almost
a copy-paste of [how tests create freeze
files](https://github.com/PennyLaneAI/pennylane/blob/f7bba2b8a21319309b21250073492573f5ee967a/.github/workflows/tests.yml#L42-L91)
on master today, but with some tweaks.

**Description of the Change:**
Run `pip freeze` right before building the docs, then push it to (or
push and open) a PR to maintain the stable results on GitHub. Some
notes:
- There was a missing `&&` so this action is not currently installing
proper environments. I fixed it here
- unlike tests, I save the results to a file with the `.tmp` suffix.
This is because I have to create it using the `pre-build-command` on the
user's branch, whereas I am able to check out the deps branch before on
the test suite. Subtle difference, but I had conflicts without it.
Praise be `mv -f`

For example, check out the commit made to an existing PR
[here](https://github.com/PennyLaneAI/pennylane/pull/5210/commits/77428e0fafe99ce3decf8bc0bf416186bdcca1bc)

**Benefits:**
- We will have a record of stable dependencies for building docs
- The docs environment is now correctly configured for this check

**Possible Drawbacks:**
- this runs on PRs, and we didn't want that to happen for tests. I'm
only like 70% sure that there won't be unexpected errors because of it
- I was considering removing this action in #5136 altogether, and only
having one docs build run (yep, we build it twice!). the context for why
is at the top of `docs.yml`, but I was hoping to fix it. Unfortunately
(as you can see in that PR), I can't make this change within the RTD
build because it doesn't have write-access to GitHub. If anyone has a
clue of how to do it there instead (maybe rsync to somewhere else, then
have that action commit the change?), please lmk
- if two PRs run this simultaneously, there might be a race condition,
and that would probably make the latter PR's CI fail due to the branch
being out-of-date with the remote. they'd have to finish the
"Documentation Check" action within a few seconds of each other AND
there be an outdated dependency for this to occur, so it seems rather
unlikely

---------

Co-authored-by: Mikhail Andrenkov <mikhail@xanadu.ai>