[Bugfix] Make sure Identity, X, Y, Z have correct __name__ by default (#5252)
Before, the overwriting of the class name happens in the initialization,
which means that it is not overwritten until at least one class instance
is initialized. This lead to problems in datasets. Fixed this by
globally overwriting the class name instead.

Better long term solution:
Check for instances instead of class names (this is a bigger effort
though as this happens in a lot of places in the codebase and, likely,
plugins)