run pylint in format action (#5227)
the current pylint action uses
[ricardochaves/python-lint](https://github.com/ricardochaves/python-lint)
which is unmaintained. I searched around, and realized there are no
actively-maintained pylint actions! This is a concern for one main
reason: that action is dockerized and only runs with python 3.9.
Unfortunately, `pylint` gives different results with different python
versions. One day, we might not be able to lint pennylane using python
3.9 (but we probably could - it's not like `pylint` needs to run any
code!), and we'd be stuck!

I started linting some files using the latest version of `pylint`, but I
realized that there are way more files with issues than I thought so I
just pinned it to the current version. We can start addressing those
later. ~Lmk if you want me to revert the changes and just deal with it
all later.~ EDIT: I have reverted the changes.

To be clear, there are 2 different things that will need upgrading (not
in this PR, but later):
1. switching to the latest version of pylint
2. linting with a newer version of Python

PS: I tried to just use `pylint pennylane/**/*.py` as
[suggested](https://github.com/pylint-dev/pylint/pull/2758#issuecomment-561164761)
by a potential pylint contributor, but found that `**` evaluates
differently in different shells (eg. old bash vs. new bash vs. zsh) and
didn't do what I wanted on the `ubuntu-latest` runner

PPS: I will need someone with access to change the branch protection to
accommodate however this ends up