Apply all user transforms before drawing (#5277)
Fixes #5275 [sc-57775]

Basically, we were only applying the transforms before drawing when the
user provided `expansion_strategy` wasn't `"device"` and the device was
from the new device interface. This PR makes it so that we always apply
all the user transforms before drawing. This is a quick fix until #5139
can get in.