Upgrade null.qubit to the new device API (#5211)
**Context:**
null.qubit doesn't quite do everything as it should, so it needs an
upgrade of sorts.

**Description of the Change:**
Port null.qubit to the new Device API. The diff for the device file can
be tough to parse, so let me share some context for what I did. In the
fist commit, I just copy-pasted the contents of default_qubit.py, then
replaced all actual computation methods with dummy methods (eg.
`_simulate()`, `_vjp()`, etc.). It's probably best to filter out that
commit using the github UI when reviewing the changes. I'd even
recommend reviewing commits 2+3 together, then commits 4-end together as
well, just so it's clear to see what I've actually done.

As well, I defined `ClassicalShadow.shape` according to what it returns
with default.qubit. While testing, I found that it ignores batching
entirely. We might want to raise an error for that, but for now I'm just
raising the error in null.qubit when it tries to go down that road.

**Benefits:**
null.qubit does what it's supposed to!

**Possible Drawbacks:**
`NullQubit.compute_vjp` doesn't work in all cases, particularly with
`jax`. this can be improved in the future, but the reason appears to be
that jax is doing some smart stuff behind the scenes with inputs and
outputs that I'm not anticipating

[sc-51495]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>