`generator` methods of operators should return new opmath instances (#5253)
**Context:** Some operators have a generator method. We want to ensure
they return the new opmath instances (`SProd,` `Prod,` `Sum`). In
particular, no `Tensor` or `Hamiltonian` instances exist.

**Description of the Change:** All generators in the code have been
modified, except those in the `qchem_ops` module.
If we modify them, we will have problems with the associated tests since
`tapering` is still not fully compatible with the new opmath. Such
generators will be replaced in the upcoming future.

**Benefits:** Modifying the generators is necessary to enable the new
opmath by default.

**Possible Drawbacks:** Since the returned generators are instances of a
different class, drawbacks can arise if we try to use `Tensor` or
`Hamiltonian` methods for the 'new' generators not implemented in the
new opmath. Similarly, we get unexpected behavior if the methods are the
same but return the results differently.

**Related GitHub Issues:** None

[sc-54079]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>