Implementing function to convert opmath instances to Hamiltonian (#5309)
**Context:** To enable the new opmath by default, we want to implement a
function to transform an opmath instance into a `Hamiltonian.`

**Description of the Change:** A new function `convert_to_legacy_H` has
been implemented to convert opmath instances (`Sum,` `SProd,` and
`Prod`) to `Hamiltonian` instances.

**Benefits:** Besides being useful for some internal functionality, this
implementation completes the transformation from a `Hamiltonian` to an
opmath instance with the corresponding inverse operation.

**Possible Drawbacks:** The opmath instance to transform is initially
simplified using the `qml.simplify()` function. Therefore, the final
output depends on the behavior of such a function. Furthermore, if the
operator is not an opmath instance, it is returned directly.

**Related GitHub Issues:** None.

[sc-57909]

---------

Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>