Add grouping functionality to `Sum` (#5179)
**Context:**
This PR adds support for grouping with `Sum`.

**Description of the Change:**
* Added `grouping_type` and `method` arguments to `qml.dot`, `qml.sum`,
and `qml.ops.Sum` constructor.
* Added `Sum.grouping_indices` attribute which stores grouping indices
in the same manner as `qml.Hamiltonian`.
* Added `Sum.compute_grouping` which updates `Sum.grouping_indices`
based on the requested grouping type and method. The grouping indices
refer to the list of ops returned by `Sum.terms`, not `Sum.operands`. I
added a note to the docstring of the function, class, and constructor
functions about this.

**Benefits:**
One step closer to `Hamiltonian`-`Sum` feature parity.

**Possible Drawbacks:**

**Related GitHub Issues:**


TODO: Add changelog entry with example.

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: qottmann <korbinian.kottmann@gmail.com>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>