Update message from dependency bot (#5327)
As detailed in the description of #5254 , the PennyLane lightning
version in the dependency files as maintained by the bot is expected to
be a version ahead of release.

Not knowing about #5254, that confused me, and the [Update stable
dependency files](https://github.com/PennyLaneAI/pennylane/pull/5297) PR
itself doesn't provide an explanation. This should add a note in the
description of the PR the bot opens explaining that this is expected
behaviour.