Update measurement tests for new op math (#5368)
Only one test needed updating. Data type for eigvals is `float` for
`Prod` and `int` for `Tensor`. This is acceptable because what we
actually care about here is that the `sample(obs).numeric_type` matches
the `obs.eigvals` numeric type. In light of that, the test is updated to
check that instead.

In doing that, we discovered that while the numeric type for Identity is
`int` (as expected by the previous implementation of the test), the
eigvals on `Identity` are floats, causing the issue #5369. **This
particular test currently does not pass, but that is because it reveals
a bug, not because the test is wrong.**

The bug is general and not related to the `new_op_math` project, and
will be fixed on master (pending decision about how to resolve). The one
test failure in `test_sample.py` will be resolved when that fix is
merged into here.