Define `Hamiltonian.pauli_rep` (#5377)
In PR #5362 , I wanted to make `ApproxTimeEvolution` depend on the
`pauli_rep` of the provided operator. Unfortunately, we never added to
`Hamiltonian` because we didn't see the point in improving
`Hamiltonian`. But not we are in the place where it needs to have
similar behaviour to its replacement. So let's just add the `pauli_rep`
property in.

[sc-58834]

---------

Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>