Dynamic circuit support for old device API (#5317)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
The Lightning simulators derive from `QubitDevice`. With the recent
addition of `MidMeasureMP` support in `lightning.qubit`, we can
potentially enable dynamic circuit execution in the Lightning family.

**Description of the Change:**
If the tape has mid-circuit measurements before apply, pass a
`mid_measurements` dictionary through `apply` via `kwargs`. Finally
return `mid_measurements` along with the results if operating in dynamic
mode. Also add `dynamic_one_shot` to the `full_transform_program` if
MCMs are present in the tape and the device has capability
`supports_mid_measure`.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Nathan Killoran <co9olguy@users.noreply.github.com>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Mikhail Andrenkov <mikhail@xanadu.ai>