Raise error when applying `Snapshot` with measurements on `default.mixed` (#5416)
**Context:**
`default.mixed` device does not support `Snapshot` with measurements. It
returns the density matrix, which is incorrect.

**Description of the Change:**
Adds an error message to when `Snapshot` with measurements is used on
`default.mixed`.

**Benefits:**
Less confusion.

**Possible Drawbacks:**
New error raised. This is a near-term fix until we improve
`default.mixed` to properly support `Snapshot`

**Related GitHub Issues:**
https://github.com/PennyLaneAI/pennylane/issues/5202
[sc-56793]