Add support for error tracking with `TrotterProduct` template  (#5384)
**Context:**
Adding support to compute the error for a given instance of
`qml.TrotterProduct`. This is part of the push on error propagation in
Pennylane

**Description of the Change:**
Using the paper ["Theory of Trotter Error with Commutator
Scaling"](https://arxiv.org/pdf/1912.08854.pdf) to implement
upper-bounds on the additive error for the Suzuki Trotter Product
formula.

- Created an error module: `pennylane/resources/error`, moved the base
`error.py` file into this repository (similar changes for the tests)
- Created a `trotter_error.py` file in the error module which contains
all of the helper functions along with the implementations for the
`_one_norm_error` and `_commutator_error` functions.
- Updated the abstract `ErrorOperation` to change the `error()` from a
property to a method. (associated changes in tests)
- Updated `TrotterProduct` to inherit from `ErrorOperation` and
implemented the `error()` method.
- Added integration tests for the new error method in the TrotterProduct
class.

Using this error method, we can benchmark the error in approximating
time evolution using the Suzuki-Trotter product formulas:


![image](https://github.com/PennyLaneAI/pennylane/assets/21964864/a589e4da-37fd-4cb3-93e3-067f2668103f)

---------

Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>