Update qml.generator's "hamiltonian" method to follow new operator arithmetic (#5421)
**Context:**
We are generally treating "Hamiltonian" as meaning alternately
`qml.ops.Hamiltonian` or `qml.ops.LinearCombination` depending on
whether or not `new_opmath` is enabled.

The current implementation of `qml.generator(op, format="hamiltonian")`
mostly respects that, but uses `convert_to_legacy_H` if the generator is
a `Sum`, `Prod` or `SProd` even with `new_opmath` is enabled, breaking
the convention.

**Description of the Change:**
We add `convert_to_H`, which converts any arithmetic operator to a
conceptual Hamiltonian - either `qml.ops.Hamiltonian` or
`qml.ops.LinearCombination`. We use the new `convert_to_H` function in
the `qml.generator` method. The `convert_to_legacy_H` function continues
to convert specifically to a legacy Hamiltonian regardless of the global
setting for `new_opmath`.

**Benefits:**
Consistency, no one will ask for a generator with `new_opmath` enabled
and get a legacy Hamiltonian.

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: qottmann <korbinian.kottmann@gmail.com>
Co-authored-by: Alex Preciado <alex.preciado@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>